+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.4 (2017-09-08)
+------------------
+* Add shebang selection from different versions of the same interpreter, or different interpreters associated with the same extension
+* Generalize exception handling
+* Reading mappings (`{extension: interpreter}`) from a `json` file as a dict (using `eval` (will be changed to be `json` parsing))
+* Enhance the `shebang` function
+
 0.1.3 (2017-09-02)
 ------------------
 * Enhance some aspects